requirements.txt
flask==2.3.3
flask-cors==4.0.0
Pillow==10.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pdf2image==1.16.3 